veridoc-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
google-generativeai==0.3.1
pytesseract==0.3.10
pdf2image==1.16.3